requirements.txt
fastapi
httpx
pydantic
pydantic-settings
PyJWT
pytest
uvicorn